refactor(potree2): revert worker import to work as expected
