+++ .babelrc
@@ -22,6 +22,8 @@
             "regenerator": false
         }],
         ["@babel/plugin-proposal-nullish-coalescing-operator"],
+        ["@babel/plugin-transform-runtime"],
+        ["babel-plugin-transform-import-meta"],
         ["transform-jsbi-to-bigint"],
         ["minify-replace", {
             "replacements": [{
